ZCS-8785:Fix for issue with data type range values for Mysql

+++ store/src/java/com/zimbra/cs/imap/ImapSessionManager.java
@@ -772,7 +772,7 @@ final class ImapSessionManager {
     private String cacheKey(FolderStore folder, boolean active) {
         MailboxStore mbox = folder.getMailboxStore();
         int modseq = folder instanceof SearchFolderStore ? mbox.getLastChangeID() : folder.getImapMODSEQ();
-        int uvv = folder instanceof SearchFolderStore ? mbox.getLastChangeID() : ImapFolder.getUIDValidity(folder);
+        long uvv = folder instanceof SearchFolderStore ? mbox.getLastChangeID() : ImapFolder.getUIDValidity(folder);
         String acctId = null;
         try {
             acctId = mbox.getAccountId();
